feat: add size in context
