+++ packages/blade/src/components/Accordion/index.ts
@@ -1,4 +1,4 @@
 export { Accordion } from './Accordion';
-export type { AccordionProps } from './Accordion';
+export type { AccordionProps } from './types';
 
 export * from './AccordionItem';
